requirements.txt
cffi==1.17.1
pillow==10.4.0
psutil==6.0.0
pycparser==2.22
pyheif==0.8.0
